Admin Panel: Create guide help button for individual pages
There is now a greyed-out guide help button available in the top panel 
of the Admin Panel. On screen load it checks if there is a help text 
associated with this page (it checks the URL, but if there is no match, 
then it searches by title). If it is found, the button changes to a more 
visible color. Pressing the button brings a guide help screen, which 
explains useful information pertaining to this particular Admin Panel 
screen.